Fix: Detect the existence of patch files

+++ Localization.py
@@ -211,6 +211,8 @@ def set_language(lang, force=False):
                                              ("Context.sublime-menu.txt", CONTEXT)):
                 PATCH_RES = "Packages/{}/patch/{}/{}/{}".format(
                     PACKAGE_NAME, patch_version, lang, patch_file_name)
+                if not os.path.isfile(PATCH_RES):
+                    continue
                 content = sublime.load_binary_resource(PATCH_RES)
                 with open(org_file_name, 'wb') as f:
                     f.write(content)
